+++ model_compression_toolkit/qat/keras/quantizer/ste_rounding/symmetric_ste.py
@@ -18,7 +18,9 @@ from typing import Union
 import numpy as np
 import tensorflow as tf
 from tensorflow.python.framework.tensor_shape import TensorShape
-from model_compression_toolkit.core.common.constants import SIGNED, FQ_MIN, FQ_MAX
+from model_compression_toolkit.core.common.constants import SIGNED
+from model_compression_toolkit.quantizers_infrastructure.constants import FQ_MIN, FQ_MAX
+
 from model_compression_toolkit.qat import TrainingMethod
 
 from model_compression_toolkit.target_platform_capabilities.target_platform import QuantizationMethod
